FAI-227 - Add mapping to Viewmodel from Vacancies response

+++ src/SFA.DAS.Campaign.Application/Vacancies/Queries/GetVacancies/GetVacanciesQueryResult.cs
@@ -5,7 +5,7 @@ namespace SFA.DAS.Campaign.Application.Vacancies.Queries.GetVacancies
 {
     public class GetVacanciesQueryResult
     {
-        public long TotalFound { get ; set ; }
+        public int TotalFound { get ; set ; }
         public List<Vacancy> Vacancies { get ; set ; }
         public Location Location { get ; set ; }
         public List<Route> Routes { get ; set ; }
